add process_cocodata
process_cocodata 用于将提取coco2017数据集的信息并输出到txt
